backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.0
pymysql==1.1.0
python-dotenv==1.0.0
passlib==1.7.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
openai>=1.12.0
# Naver Clova STT 관련
aiofiles==23.2.1
# 오디오 처리
pydub==0.25.1
# 이메일 검증
email-validator==2.1.0
# MySQL 암호화
cryptography==41.0.7
# 비밀번호 해싱
bcrypt==4.1.2
# FastAPI 데이터 검증
pydantic==2.5.0
# FastAPI 보안
fastapi-security==0.1.0 